Added Checks for Neighboring Nodes
When a neighboring node is already in the open list (has been discovered through another node), we check to see if taking a path through the current node to the neighboring node would produce a better path. If so, we update the neighbor's parent to be the current node and update its g-value.